requirements.txt
crewai
litellm
python-dotenv
requests
flask
markdown
Pygments
